refactoring - read_template function
